Upgrade to version 7.2 of the boehm gc.
Estimated hours taken: 4
Branches: main

Upgrade to version 7.2 of the boehm gc.

This upgrade was done as 7.2 alpha 4 doesn't compile cleanly under Solaris 10.

+++ boehm_gc/libatomic_ops/src/atomic_ops/sysdeps/gcc/alpha.h
@@ -27,7 +27,6 @@ AO_nop_full(void)
 {
   __asm__ __volatile__("mb" : : : "memory");
 }
-
 #define AO_HAVE_nop_full
 
 AO_INLINE void
@@ -35,7 +34,6 @@ AO_nop_write(void)
 {
   __asm__ __volatile__("wmb" : : : "memory");
 }
-
 #define AO_HAVE_nop_write
 
 /* mb should be used for AO_nop_read().  That's the default.    */
@@ -60,7 +58,6 @@ AO_compare_and_swap(volatile AO_t *addr,
                      :"=&r" (temp), "=m" (*addr), "=&r" (was_equal)
                      : "r" (new_val), "Ir" (old)
                      :"memory");
-  return was_equal;
+  return (int)was_equal;
 }
-
 #define AO_HAVE_compare_and_swap
